Changes Launcher to accept the actual name of the tcz as an argument.
Whenever de application being executed is named "Launcher", it checks for additional arguments and uses the first one as the name of the tcz to start the application. This will make testing a lot easier.

I'm still learning about vscode , so I'm not sure if I should commit launch.json, but I'm doing it anyway.